requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
dnspython==2.3.0
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
multidict==6.0.4
openai==0.27.2
pymongo==4.3.3
python-dotenv==1.0.0
python-telegram-bot==20.1
rfc3986==1.5.0
tqdm==4.65.0
yarl==1.8.2
